Workout_Vision/requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
fastapi==0.104.0
uvicorn==0.24.0
pydantic==2.5.2
ultralytics==8.3.87
openai==1.12.0
python-multipart==0.0.6
torch==2.1.1
python-dotenv==1.0.0 
httpx==0.27.2
supabase>=2.4.0